mbot/pirate/requirements.txt
# Telegram Bot
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Media Downloaders
yt-dlp>=2023.3.4
instaloader>=4.10
praw>=7.7.0
pytube>=15.0.0
spotipy>=2.23.0

# Utils
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.5
aiofiles>=23.1.0
tqdm>=4.66.1
validators>=0.20.0
psutil>=5.9.0